requirements.txt
asgiref==3.7.2
Django==4.2.6
django-ckeditor==6.7.0
django-cleanup==8.0.0
django-js-asset==2.1.0
django-phonenumber-field==7.2.0
django-resized==1.0.2
djangorestframework==3.14.0
phonenumbers==8.13.22
Pillow==10.0.1
pytz==2023.3.post1
sqlparse==0.4.4
